[#2933] Move ConnectionFactory related classes
Moved ConnectionFactory and its implementation to amqp-connection
module. This is a first step in removing the Proton-J dependencies from
the core module.
