+++ clients/amqp-connection/src/main/java/org/eclipse/hono/client/amqp/connection/ConnectTimeoutException.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2016, 2019 Contributors to the Eclipse Foundation
+ * Copyright (c) 2016, 2022 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
 
-package org.eclipse.hono.connection;
+package org.eclipse.hono.client.amqp.connection;
 
 /**
  * An exception thrown when a connection attempt has timed out.
