+++ service-base-quarkus/pom.xml
@@ -39,6 +39,11 @@
       <artifactId>hono-service-base</artifactId>
     </dependency>
     <dependency>
+     <groupId>org.eclipse.hono</groupId>
+     <artifactId>hono-client-amqp-connection</artifactId>
+     <optional>true</optional>
+    </dependency>
+    <dependency>
       <groupId>io.opentracing</groupId>
       <artifactId>opentracing-util</artifactId>
     </dependency>
