+++ service-base/pom.xml
@@ -1,5 +1,5 @@
 <!--
-    Copyright (c) 2016, 2021 Contributors to the Eclipse Foundation
+    Copyright (c) 2016, 2022 Contributors to the Eclipse Foundation
    
     See the NOTICE file(s) distributed with this work for additional
     information regarding copyright ownership.
@@ -32,6 +32,11 @@
     <dependency>
      <groupId>org.eclipse.hono</groupId>
      <artifactId>hono-client</artifactId>
+    </dependency>
+    <dependency>
+     <groupId>org.eclipse.hono</groupId>
+     <artifactId>hono-client-amqp-connection</artifactId>
+     <optional>true</optional>
     </dependency>
     <dependency>
       <groupId>org.eclipse.hono</groupId>
